refactor(jazz): organize query language responsibilities
